+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/matchers.kt
@@ -3,6 +3,9 @@ package io.kotlintest.matchers
 import io.kotlintest.Matcher
 import io.kotlintest.Result
 
+fun Any.shouldHaveSameHashCodeAs(other: Any) = this should haveSameHashCodeAs(other)
+fun Any.shouldNotHaveSameHashCodeAs(other: Any) = this shouldNot haveSameHashCodeAs(other)
+
 fun haveSameHashCodeAs(other: Any) = object : Matcher<Any> {
   override fun test(value: Any): Result {
     return Result(