+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/matchers/types/matchers.kt
@@ -2,6 +2,9 @@ package io.kotlintest.matchers.types
 
 import io.kotlintest.Matcher
 import io.kotlintest.Result
+import io.kotlintest.should
+
+inline fun <reified T : Annotation> Class<Any>.shouldHaveAnnotation(klass: Class<T>) = this should haveAnnotation(klass)
 
 @Suppress("UNUSED_PARAMETER")
 inline fun <A, reified T : Annotation> haveAnnotation(klass: Class<T>) = object : Matcher<Class<A>> {